Add buffalo chicken recipe
Edited index page to include a link to the buffalo chicken recipe.

The buffalo chicken page was initialized and built out to match the
other recipe pages. An image for the buffalo chicken dip was added to
the buffalo chicken page. A description, ingredients, and directions
were added to the buffalo chicken page.